requirements.txt
# Core Framework
flask
markdown

# The unified Google Generative AI SDK (for Gemini)
google-genai

# For handling environment variables
python-dotenv

# For downloading YouTube video information/transcripts
pytube

# Google authentication library
google-auth

# Pin urllib3 to a version compatible with the system's OpenSSL
urllib3<2.0
